Add files via upload
From the 4,000 terms in the Global Cryospheric Watch glossary I extracted all terms that I could identify as being related to glaciology. I then removed duplicates, giving preference to definitions by NSIDC and Molina. I discarded some archaic and esoteric terms. I removed the "glacial" adjective where the meaning of the word was clear without it. I put modifiers after root term, i.e. “Cirque glacier” was changed to “glacier, Cirque”
I then sorted the terms into the following categories:
- geologic epocs (31 terms)
- hazards (6 terms all describing the same phenomenon)
- instruments-monitoring (17 terms)
- landforms (88 terms)
- mass balance (138 terms)
- surface features (76 terms)
- uncategorized (210 terms)
My motivation for doing this is to help us arrive at a smaller set of terms that cover all important aspects of glaciers.
The categorization isn’t perfect, and more could be done with the remaining uncategorized terms (which contains 52 types of glaciers and about a dozen types of Ice Caps / Ice Sheets).
The surface features category includes all things related to snow metamorphosis, crevasses, infiltration, supraglacial lakes (lakes that are not on the glacier surface are in landforms).
The mass balance category includes everything related to adding or removing mass from a glacier, including calving. I included terms related to the delineation of zones of surface properties related to mass exchange (e.g. accumulation and ablation). 
The next step would be to review these terms and select the ones we want to work with in the workshop.