A short list of terms to start with
Ruth Duerr and Pier Luigi Buttigieg selected this short list of terms (a total of 33) as being the most relevant to the lifecycle of a glacier.  Of those, the terms: ablation, abrasion, accumulation, calving, equilibrium line, firn, glacier, ice, neve, nivation, and plucking were selected as the core terms to address by Pier.